control_room/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.5,<3.0
pydantic-settings>=2.0,<3.0
pyyaml>=6.0.1
aiofiles>=23.2
